+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * fix bug with old Aerospike version download, requiring username/password
 * add notes at the end of AMS and exporter installation, referencing one-another to avoid confusion
 * allow for aerolab to auto-create security groups in aws, as well as discover subnet IDs based on AZ for the default VPC
+* allow tagging instances and images with custom tags when using the aws backend
 
 #### 4.4.4
 * fix issue with the `autoload` directory execution in AWS